+++ docs/source/history.rst
@@ -7,10 +7,11 @@ Latest (unreleased)
 -------------------
 
 Improvements for multiple devices:
-    * Add ``scan_serial`` parameter to :class:`EasyMCP2221.Device` to prevent 
+    * Add ``scan_serial`` parameter to :class:`EasyMCP2221.Device` to prevent
       interference with other MCP2221 when searching for some Serial number.
     * Tests for corner cases in device selection with :func:`enable_cdc_serial` both enabled and disabled.
     * Improved catalog now make sure no device with the same USB system path is open twice in the same program.
+    * The catalog takes into account that the device may have been changed to another port or disconnected.
     * Fully rewrite device selection code.